NN optimizations
NeuralLayers now take activation as a parameter rather than input wire,
ensuring that the unused modules do not have to be constructed at all.

MatMulSeq now comes with further optimizations by cutting off transposition
of B and takes B_T directly as input. For NN, this is irrelevant as B is always
a vector (therefore, B = B_T in this representation).

Changed param name in Softmax module for consistency.